requirements.txt
requests
python-dotenv
azure-keyvault-secrets
azure-identity
